src/requirements.txt
motor==3.1.1
pymongo==4.3.3
cohere==5.5.8
PyMuPDF==1.24.3
fastapi==0.110.2
aiofiles==23.2.1
openai==1.35.13
langchain==0.1.20
google-generativeai
python-dotenv==1.0.1
pydantic-mongo==2.3.0
qdrant-client==1.10.1
python-multipart==0.0.9
pydantic-settings==2.2.1
uvicorn[standard]==0.29.0
